[Test] Remove ti test from taichi package.
[ghstack-poisoned]
