+++ python/taichi/main.py
@@ -666,92 +666,6 @@ class TaichiMain:
 
         return None
 
-    @staticmethod
-    def _test_python(args):
-        print("\nRunning Python tests...\n")
-
-        test_38 = sys.version_info >= (3, 8)
-
-        root_dir = ti.package_root()
-        test_dir = os.path.join(root_dir, 'tests')
-        test_dir_38 = os.path.join(root_dir, 'tests38')
-        pytest_args = []
-
-        # TODO: use pathlib to deal with suffix and stem name manipulation
-        if args.files:
-            # run individual tests
-            for f in args.files:
-                # auto-complete file names
-                if not f.startswith('test_'):
-                    f = 'test_' + f
-                if not f.endswith('.py'):
-                    f = f + '.py'
-                file = os.path.join(test_dir, f)
-                file_38 = os.path.join(test_dir_38, f)
-                has_tests = False
-                if os.path.exists(file):
-                    pytest_args.append(file)
-                    has_tests = True
-                if os.path.exists(file_38) and test_38:
-                    pytest_args.append(file_38)
-                    has_tests = True
-                assert has_tests, f"Test {f} does not exist."
-        else:
-            # run all the tests
-            pytest_args = [test_dir]
-            if test_38:
-                pytest_args += [test_dir_38]
-        if args.verbose:
-            pytest_args += ['-v']
-        if args.rerun:
-            pytest_args += ['--reruns', args.rerun]
-        try:
-            if args.coverage:
-                pytest_args += ['--cov-branch', '--cov=python/taichi']
-            if args.cov_append:
-                pytest_args += ['--cov-append']
-            if args.keys:
-                pytest_args += ['-k', args.keys]
-            if args.marks:
-                pytest_args += ['-m', args.marks]
-            if args.failed_first:
-                pytest_args += ['--failed-first']
-            if args.fail_fast:
-                pytest_args += ['--exitfirst']
-        except AttributeError:
-            pass
-
-        try:
-            from multiprocessing import cpu_count  # pylint: disable=C0415
-            threads = min(8, cpu_count())  # To prevent running out of memory
-        except NotImplementedError:
-            threads = 2
-
-        if not os.environ.get('TI_DEVICE_MEMORY_GB'):
-            os.environ['TI_DEVICE_MEMORY_GB'] = '1.0'  # Discussion: #769
-
-        env_threads = os.environ.get('TI_TEST_THREADS', '')
-        threads = args.threads or env_threads or threads
-        print(f'Starting {threads} testing thread(s)...')
-        if args.show_output:
-            pytest_args += ['-s']
-            print(
-                f'Due to how pytest-xdist is implemented, the -s option does not work with multiple thread...'
-            )
-        else:
-            if int(threads) > 1:
-                pytest_args += ['-n', str(threads)]
-        import pytest  # pylint: disable=C0415
-        return int(pytest.main(pytest_args))
-
-    @staticmethod
-    def _test_cpp(args):
-        # Cpp tests use the legacy non LLVM backend
-        ti.reset()
-        print("Running C++ tests...")
-        task = ti.Task('test')
-        return int(task.run(*args.files))
-
     @register
     def benchmark(self, arguments: list = sys.argv[2:]):
         """Run Python tests in benchmark mode"""
@@ -810,129 +724,9 @@ class TaichiMain:
 
     @register
     def test(self, arguments: list = sys.argv[2:]):
-        """Run the tests"""
-        parser = argparse.ArgumentParser(prog='ti test',
-                                         description=f"{self.test.__doc__}")
-        parser.add_argument('files',
-                            nargs='*',
-                            help='Test name(s) to be run, e.g. "cli"')
-        parser.add_argument('-c',
-                            '--cpp',
-                            dest='cpp',
-                            action='store_true',
-                            help='Only run the C++ tests')
-        parser.add_argument('-s',
-                            '--show',
-                            dest='show_output',
-                            action='store_true',
-                            help='Show output (do not capture)')
-        parser.add_argument('-v',
-                            '--verbose',
-                            dest='verbose',
-                            action='store_true',
-                            help='Run with verbose outputs')
-        parser.add_argument('-r',
-                            '--rerun',
-                            required=False,
-                            default=None,
-                            dest='rerun',
-                            type=str,
-                            help='Rerun failed tests for given times')
-        parser.add_argument('-k',
-                            '--keys',
-                            required=False,
-                            default=None,
-                            dest='keys',
-                            type=str,
-                            help='Only run tests that match the keys')
-        parser.add_argument('-m',
-                            '--marks',
-                            required=False,
-                            default=None,
-                            dest='marks',
-                            type=str,
-                            help='Only run tests with specific marks')
-        parser.add_argument('-f',
-                            '--failed-first',
-                            required=False,
-                            default=None,
-                            dest='failed_first',
-                            action='store_true',
-                            help='Run the previously failed test first')
-        parser.add_argument('-x',
-                            '--fail-fast',
-                            required=False,
-                            default=None,
-                            dest='fail_fast',
-                            action='store_true',
-                            help='Exit instantly on the first failed test')
-        parser.add_argument('-C',
-                            '--coverage',
-                            required=False,
-                            default=None,
-                            dest='coverage',
-                            action='store_true',
-                            help='Run tests and record the coverage result')
-        parser.add_argument(
-            '-A',
-            '--cov-append',
-            required=False,
-            default=None,
-            dest='cov_append',
-            action='store_true',
-            help=
-            'Append coverage result to existing one instead of overriding it')
-        parser.add_argument(
-            '-t',
-            '--threads',
-            required=False,
-            default=None,
-            dest='threads',
-            type=str,
-            help='Custom number of threads for parallel testing')
-        parser.add_argument(
-            '-a',
-            '--arch',
-            required=False,
-            default=None,
-            dest='arch',
-            type=str,
-            help='Custom the arch(s) (backend) to run tests on')
-        parser.add_argument(
-            '-n',
-            '--exclusive',
-            required=False,
-            default=False,
-            dest='exclusive',
-            action='store_true',
-            help=
-            'Exclude arch(s) from test instead of include them, together with -a'
+        raise RuntimeError(
+            'ti test is deprecated. Please run `python tests/run_tests.py` instead.'
         )
-
-        args = parser.parse_args(arguments)
-
-        if args.arch:
-            arch = args.arch
-            if args.exclusive:
-                arch = '^' + arch
-            print(f'Running on Arch={arch}')
-            os.environ['TI_WANTED_ARCHS'] = arch
-
-        # Short circuit for testing
-        if self.test_mode: return args
-
-        if args.files:
-            if args.cpp:
-                return TaichiMain._test_cpp(args)
-            return TaichiMain._test_python(args)
-        if args.cpp:
-            # Only run C++ tests
-            return TaichiMain._test_cpp(args)
-        # Run both C++ and Python tests
-        ret = TaichiMain._test_python(args)
-        if ret != 0:
-            return ret
-        return TaichiMain._test_cpp(args)
 
     @register
     def run(self, arguments: list = sys.argv[2:]):
